requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
requests==2.31.0
Werkzeug==3.0.1
python-pptx==0.6.23
